+++ h/grid.cuh
@@ -10,7 +10,7 @@ namespace PhysPeach{
         uint* refCell_dev;
 
         //cell list
-        float rc;
+        double rc;
         uint M;
         uint EpM;//(Num of Elements per M) - 1
         uint* cell_dev;//[M2][EpM] ->[m2* EpM + epm] epm = 0->NofE
@@ -20,31 +20,31 @@ namespace PhysPeach{
         //for setters and getters
         float *getNU_dev[2]; //U[N]
     };
-    void makeGrid(Grid* grid, float L);
+    void makeGrid(Grid* grid, double L);
     void killGrid(Grid* grid);
     void makeCellPattern2D(Grid* grid);
-    __global__ void updateGrid2D(Grid grid, uint* cell, float* x);
+    __global__ void updateGrid2D(Grid grid, uint* cell, double* x);
     void setUpdateFreq(Grid* grid, double dt, float *v);
-    void checkUpdate(Grid* grid, double dt, float* x, float* v);
+    void checkUpdate(Grid* grid, double dt, double* x, float* v);
     __global__ void culcHarmonicFint2D(
         Grid g, 
         uint *refCell, 
         uint *cell, 
         float *force, 
-        float L, 
+        double L, 
         float *diam, 
-        float *x);
+        double *x);
     __global__ void culcFint2D(
         Grid g,
         uint *refCell, 
         uint *cell, 
         float *force, 
-        float L, 
+        double L, 
         float *diam, 
-        float *x);
+        double *x);
     
     //for setters and getters
-    float U(Grid* grid, float *diam, float *x);
+    float U(Grid* grid, float *diam, double *x);
 }
 
 #endif
