+++ h/particles.cuh
@@ -13,11 +13,11 @@
 namespace PhysPeach{
     struct Particles{
         float *diam;
-        float *x;
+        double *x;
         float *v;
 
         float *diam_dev;
-        float *x_dev;
+        double *x_dev;
         float *v_dev;
         curandState *rndState_dev;
         float *force_dev;
@@ -28,11 +28,11 @@ namespace PhysPeach{
     };
     void makeParticles(Particles* particles);
     void killParticles(Particles* particles);
-    void scatterParticles(Particles* particles, float L);
+    void scatterParticles(Particles* particles, double L);
 
     //time evolutions
     __global__ void vEvoBD(float *v, double dt, float themalFuctor, float *force, curandState *state);
-    __global__ void xEvo(float *x, double dt, float L, float *v);
+    __global__ void xEvo(double *x, double dt, double L, float *v);
     void removevg2D(Particles* p);
 
     //setters and getters
